+++ fabric/src/main/java/io/github/sakurawald/module/initializer/command_interactive/CommandInteractiveInitializer.java
@@ -8,8 +8,24 @@ import net.minecraft.server.network.ServerPlayerEntity;
 
 import java.time.Instant;
 import java.util.BitSet;
+import java.util.HashSet;
+import java.util.Set;
 
 public class CommandInteractiveInitializer extends ModuleInitializer {
+
+    private static final Set<CommandExecutionC2SPacket> TRUSTED_PACKETS = new HashSet<>();
+
+    public static void addTrustedPacket(CommandExecutionC2SPacket packet) {
+        TRUSTED_PACKETS.add(packet);
+    }
+
+    public static void removeTrustedPacket(CommandExecutionC2SPacket packet) {
+        TRUSTED_PACKETS.remove(packet);
+    }
+
+    public static boolean isTrustedPacket(CommandExecutionC2SPacket packet) {
+        return TRUSTED_PACKETS.contains(packet);
+    }
 
     @SuppressWarnings("UnnecessaryLocalVariable")
     private static CommandExecutionC2SPacket makeCommandExecutionPacket(ServerPlayerEntity player, String commandString) {
@@ -23,7 +39,13 @@ public class CommandInteractiveInitializer extends ModuleInitializer {
     }
 
     public static void mimicCommandExecutionPacket(ServerPlayerEntity player, String commandString) {
+        /* Make command execution packet. */
+        CommandExecutionC2SPacket packet = makeCommandExecutionPacket(player, commandString);
+
+        /* Trust the packet. */
+        addTrustedPacket(packet);
+
         // NOTE: Call the method directly to simulate a c2s packet.
-        player.networkHandler.onCommandExecution(makeCommandExecutionPacket(player, commandString));
+        player.networkHandler.onCommandExecution(packet);
     }
 }